separate getImageID logic from adapter trimming functionality
