+++ slave/components/quality_controller/adapter_trimming.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"os"
 	"path"
-	"strings"
 
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
@@ -27,6 +26,7 @@ type adapterTrimming struct {
 	ctx context.Context
 }
 
+// NewAdapterTrimming function creates a adapterTrimming instance containing necessary config for adapter trimming job
 func NewAdapterTrimming(ctx context.Context, dockerCli *client.Client, config *config_parser.Config) Controller {
 	return &adapterTrimming{
 		dockerCLI: dockerCli,
@@ -35,34 +35,7 @@ func NewAdapterTrimming(ctx context.Context, dockerCli *client.Client, config *c
 	}
 }
 
-// TODO: Replace this function into a constants file where this can be used by other quality control processes
-// getImageID attempts to find the Docker image by given term
-func getImageID(client *client.Client, ctx context.Context, term string) (string, error) {
-	images, err := client.ImageList(ctx, types.ImageListOptions{})
-	if err != nil {
-		return "", fmt.Errorf("failed to get available Docker images, err: %v", err)
-	} else if len(images) == 0 {
-		return "", fmt.Errorf("getImageID found no images")
-	}
-	found := false
-	assemblerID := ""
-	for _, im := range images {
-		if found {
-			break
-		}
-		for _, tag := range im.RepoTags {
-			if strings.Contains(tag, term) {
-				found = true
-				assemblerID = im.ID
-			}
-		}
-	}
-	if !found {
-		return "", fmt.Errorf("failed to find a Docker container for the given assembler")
-	}
-	return assemblerID, nil
-}
-
+// Process does the adapter trimming on raw input data
 func (a *adapterTrimming) Process() (string, error) {
 
 	// TrimmedFileName is the filename where the trimmed reads are going to be stored.
