+++ slave/components/quality_controller/utils.go
@@ -1,0 +1,37 @@
+package quality_controller
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/client"
+)
+
+// getImageID attempts to find the Docker image by given term
+func getImageID(client *client.Client, ctx context.Context, term string) (string, error) {
+	images, err := client.ImageList(ctx, types.ImageListOptions{})
+	if err != nil {
+		return "", fmt.Errorf("failed to get available Docker images, err: %v", err)
+	} else if len(images) == 0 {
+		return "", fmt.Errorf("getImageID found no images")
+	}
+	found := false
+	assemblerID := ""
+	for _, im := range images {
+		if found {
+			break
+		}
+		for _, tag := range im.RepoTags {
+			if strings.Contains(tag, term) {
+				found = true
+				assemblerID = im.ID
+			}
+		}
+	}
+	if !found {
+		return "", fmt.Errorf("failed to find a Docker container for the given assembler")
+	}
+	return assemblerID, nil
+}
